refactor: Organize modules in backend
