+++ backend/rag_modules/google_genai_database.py
@@ -9,8 +9,9 @@ import os
 
 def connect_db(db_name=None):
     if db_name is None:
-        script_dir = os.path.dirname(os.path.abspath(__file__))
-        db_name = os.path.join(script_dir, "vector-store", "genai_embeddings.db")
+        # Set the database path to the backend root directory
+        backend_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+        db_name = os.path.join(backend_root, "vector-store", "genai_embeddings.db")
     # Ensure the directory exists
     os.makedirs(os.path.dirname(db_name), exist_ok=True)
     return sqlite3.connect(db_name)
